Finished Xylophone App development
Finally, finished this Xylophone App development from Angela Yu App Brewery's iOS course

This app now run perfectly as it should be and designed in mind for iPhone 11. I might be added Auto Layout and Responsive UI to this app (also to the Dicee and Magic 8 Ball app) so i can sideload these app to test it natively on my iPhone SE (2020).